requirement.txt
django==5.2.6
django-cors-headers==4.7.0
python-dotenv==1.0.0
requests==2.32.5
celery==5.3.4
django-celery-beat==2.8.1
redis==5.0.1
flower==2.0.1
PyMySQL==1.1.1
cryptography==43.0.3
gunicorn==22.0.0
corsheaders